# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Server.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/candidate": {
      "post": {
        "tags": [
          "Candidate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Candidate"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/candidate/{id}": {
      "put": {
        "tags": [
          "Candidate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCandidateCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditCandidateCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditCandidateCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Candidate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Candidate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/candidate/verify-bg/{id}": {
      "put": {
        "tags": [
          "Candidate"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/bulk/exel-doc": {
      "post": {
        "tags": [
          "Candidate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidatesWithExelDocCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidatesWithExelDocCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidatesWithExelDocCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/resume-doc": {
      "post": {
        "tags": [
          "Candidate"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateWithResumeDocCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateWithResumeDocCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCandidateWithResumeDocCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/interview": {
      "post": {
        "tags": [
          "Interview"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/interview/{id}": {
      "put": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditInterviewCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditInterviewCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditInterviewCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/interview/{id}/feedback": {
      "post": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewFeedbackCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewFeedbackCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewFeedbackCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/interview/{interviewId}/feedback/{feedbackId}": {
      "put": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "interviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "feedbackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditInterviewFeedbackCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EditInterviewFeedbackCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EditInterviewFeedbackCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "interviewId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "feedbackId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/interview/{id}/move-status": {
      "put": {
        "tags": [
          "Interview"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveInterviewStatusCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveInterviewStatusCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MoveInterviewStatusCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/job-application/{id}": {
      "delete": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CandidateSkillDTO": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "CreateCandidateCommand": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "dob": {
            "type": "string",
            "format": "date-time"
          },
          "collegeName": {
            "type": "string",
            "nullable": true
          },
          "resumeUrl": {
            "type": "string",
            "nullable": true
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateSkillDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateCandidateWithResumeDocCommand": {
        "type": "object",
        "additionalProperties": false
      },
      "CreateCandidatesWithExelDocCommand": {
        "type": "object",
        "additionalProperties": false
      },
...
      "CreateInterviewCommand": {
        "type": "object",
        "properties": {
          "jobApplicationId": {
            "type": "string",
            "format": "uuid"
          },
          "roundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "interviewType": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationInMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "meetingLink": {
            "type": "string",
            "nullable": true
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewParticipantDTO"
            },
            "nullable": true,
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "CreateInterviewFeedbackCommand": {
        "type": "object",
        "properties": {
          "interviewId": {
            "type": "string",
            "format": "uuid"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "skillFeedbacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewSkillFeedbackDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DocumentDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "documentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EditCandidateCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "contactNumber": {
            "type": "string",
            "nullable": true
          },
          "dob": {
            "type": "string",
            "format": "date-time"
          },
          "collegeName": {
            "type": "string",
            "nullable": true
          },
          "resumeUrl": {
            "type": "string",
            "nullable": true
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CandidateSkillDTO"
            },
            "nullable": true
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EditInterviewCommand": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "roundNumber": {
            "type": "integer",
            "format": "int32"
          },
          "interviewType": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationInMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "meetingLink": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/InterviewStatus"
          },
          "participants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewParticipantDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "EditInterviewFeedbackCommand": {
        "type": "object",
        "properties": {
          "interviewId": {
            "type": "string",
            "format": "uuid"
          },
          "feedbackId": {
            "type": "string",
            "format": "uuid"
          },
          "comment": {
            "type": "string",
            "nullable": true
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "skillFeedbacks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewSkillFeedbackDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Gender": {
        "enum": [
          "Male",
          "Female",
          "Other",
          "PreferNotToSay"
        ],
        "type": "string"
      },
...
      "InterviewParticipantDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/InterviewParticipantRole"
          }
        },
        "additionalProperties": false
      },
      "InterviewParticipantRole": {
        "enum": [
          "Interviewer",
          "Observer",
          "NoteTaker",
          "HRRepresentative",
          "HiringManager"
        ],
        "type": "string"
      },
...
      "InterviewSkillFeedbackDTO": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "assessedExpYears": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "InterviewStatus": {
        "enum": [
          "NotScheduled",
          "Scheduled",
          "Completed"
        ],
        "type": "string"
      },
      "InterviewType": {
        "enum": [
          "Technical",
          "HR",
          "OnlineTest"
        ],
        "type": "string"
      },
...
      "MoveInterviewStatusCommand": {
        "type": "object",
        "properties": {
          "interviewId": {
            "type": "string",
            "format": "uuid"
          },
          "moveTo": {
            "$ref": "#/components/schemas/InterviewStatus"
          },
          "meetingLink": {
            "type": "string",
            "nullable": true
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}